You can now override the default server host (0.0.0.0) by providing a :host: config option in .ginatra/config.yml.
Made all the URLs generated in the application relative and put a base tag in layout.erb.

There's now a :prefix: config option (that defaults to "/"), that is used at the end of the base tag's href value.  This gives you the ability to have the now-relative Ginatra URLs be based off of something besides "/".